refactor: 코드 리펙토링
- 사용하지 않는 코드 삭제
- 퀴즈 상태 전체 끄기 메소드 위치 이동 및 쿼리 최적화
- Transactional 삭제
