+++ src/main/java/org/cotato/csquiz/api/socket/controller/SocketController.java
@@ -7,8 +7,7 @@ import org.cotato.csquiz.api.socket.dto.EducationCloseRequest;
 import org.cotato.csquiz.api.socket.dto.QuizOpenRequest;
 import org.cotato.csquiz.api.socket.dto.QuizSocketRequest;
 import org.cotato.csquiz.api.socket.dto.SocketTokenDto;
-import org.cotato.csquiz.domain.education.service.RecordService;
-import org.cotato.csquiz.domain.education.service.SocketService;
+import org.cotato.csquiz.domain.education.service.*;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -25,43 +24,45 @@ import org.springframework.web.bind.annotation.RestController;
 public class SocketController {
 
     private final SocketService socketService;
+    private final EducationService educationService;
     private final RecordService recordService;
+    private final QuizSolveService quizSolveService;
 
     @PatchMapping("/start/csquiz")
     public ResponseEntity<Void> openCSQuiz(@RequestBody @Valid QuizOpenRequest request) {
-        socketService.openCSQuiz(request);
+        educationService.openCSQuiz(request);
         recordService.saveAnswersToCache(request);
         return ResponseEntity.noContent().build();
     }
 
     @PatchMapping("/access")
     public ResponseEntity<Void> accessQuiz(@RequestBody @Valid QuizSocketRequest request) {
-        socketService.accessQuiz(request);
+        quizSolveService.accessQuiz(request);
         recordService.saveAnswer(request);
         return ResponseEntity.noContent().build();
     }
 
     @PatchMapping("/start")
     public ResponseEntity<Void> startQuizSolve(@RequestBody @Valid QuizSocketRequest request) {
-        socketService.startQuizSolve(request);
+        quizSolveService.startQuizSolve(request);
         return ResponseEntity.noContent().build();
     }
 
     @PatchMapping("/deny")
     public ResponseEntity<Void> denyQuiz(@RequestBody @Valid QuizSocketRequest request) {
-        socketService.denyQuiz(request);
+        quizSolveService.denyQuiz(request);
         return ResponseEntity.noContent().build();
     }
 
     @PatchMapping("/stop")
     public ResponseEntity<Void> stopQuizSolve(@RequestBody @Valid QuizSocketRequest request) {
-        socketService.stopQuizSolve(request);
+        quizSolveService.stopQuizSolve(request);
         return ResponseEntity.noContent().build();
     }
 
     @PatchMapping("/close/csquiz")
-    public ResponseEntity<Void> stopAllQuiz(@RequestBody @Valid EducationCloseRequest request) {
-        socketService.stopAllQuiz(request);
+    public ResponseEntity<Void> closeEducation(@RequestBody @Valid EducationCloseRequest request) {
+        educationService.closeEducation(request);
         return ResponseEntity.noContent().build();
     }
